[BENCH-98] Add DEBUG symbols and support properly; yank old cruft preventing it. Fixed DEBUG compilation bugs, bad code that had been there. Update documentation to reflect how to compile with debug symbols.
[BENCH-96] Add explicit path support for the run script
so that peer scripts are picked up automatically so long
as they are in the same directory as it. Updated documentation
so that it indicates more specifically how to set an
install path for the project, the install path can be
anywhere.

Added command line argument support for DB host name,
port number, schema name, etc. Format the connection
string in the usual schema@hostname:port syntax. Made
sure that the database name buffer is same size as the
connection string (C v C++ code) so that the C code
doesn't pass in something the C++ code cannot handle.